rollback changes to not handle api 24+ separaterly for URI schemes depending the frame where is executed
