+++ app/src/main/java/com/duckduckgo/app/browser/di/BrowserModule.kt
@@ -48,7 +48,6 @@ import com.duckduckgo.app.tabs.ui.GridViewColumnCalculator
 import com.duckduckgo.app.trackerdetection.TrackerDetector
 import dagger.Module
 import dagger.Provides
-import javax.inject.Named
 import javax.inject.Singleton
 
 @Module
@@ -65,20 +64,18 @@ class BrowserModule {
 
     @Provides
     fun browserWebViewClient(
+        requestRewriter: RequestRewriter,
         specialUrlDetector: SpecialUrlDetector,
         requestInterceptor: RequestInterceptor,
         offlinePixelCountDataStore: OfflinePixelCountDataStore,
-        uncaughtExceptionRepository: UncaughtExceptionRepository,
-        @Named(SpecialUrlHandler.MAIN_FRAME_HANDLER) mainFrameUrlHandler: SpecialUrlHandler,
-        @Named(SpecialUrlHandler.SUB_FRAME_HANDLER) subFrameUrlHandler: SpecialUrlHandler
+        uncaughtExceptionRepository: UncaughtExceptionRepository
     ): BrowserWebViewClient {
         return BrowserWebViewClient(
+            requestRewriter,
             specialUrlDetector,
             requestInterceptor,
             offlinePixelCountDataStore,
-            uncaughtExceptionRepository,
-            mainFrameUrlHandler,
-            subFrameUrlHandler
+            uncaughtExceptionRepository
         )
     }
 
@@ -162,19 +159,5 @@ class BrowserModule {
     @Provides
     fun webViewPreviewGenerator(): WebViewPreviewGenerator {
         return FileBasedWebViewPreviewGenerator()
-    }
-
-    @Provides
-    @Singleton
-    @Named(SpecialUrlHandler.MAIN_FRAME_HANDLER)
-    fun mainFrameUrlHandler(requestRewriter: RequestRewriter): SpecialUrlHandler {
-        return MainFrameUrlHandler(requestRewriter)
-    }
-
-    @Provides
-    @Singleton
-    @Named(SpecialUrlHandler.SUB_FRAME_HANDLER)
-    fun nonMainFrameUrlHandler(): SpecialUrlHandler {
-        return SubFrameUrlHandler()
     }
 }
